bindings: Move string-related bindings code to script_bindings.
