+++ components/script_bindings/Cargo.toml
@@ -21,6 +21,18 @@ phf_shared = "0.11"
 serde_json = { workspace = true }
 
 [dependencies]
+cssparser = { workspace = true }
+html5ever = { workspace = true }
+js = { workspace = true }
+jstraceable_derive = { path = "../jstraceable_derive" }
+libc = { workspace = true }
+log = { workspace = true }
+malloc_size_of = { workspace = true }
+malloc_size_of_derive = { workspace = true }
+num-traits = { workspace = true }
+regex = { workspace = true }
+servo_atoms = { workspace = true }
+servo_config = { path = "../config" }
 style = { workspace = true }
 
 [features]
