+++ components/script_bindings/lib.rs
@@ -1,3 +1,19 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#[macro_use]
+extern crate jstraceable_derive;
+#[macro_use]
+extern crate log;
+#[macro_use]
+extern crate malloc_size_of_derive;
+
+pub mod conversions;
+pub mod str;
+mod trace;
+
+// These trait exports are public, because they are used in the DOM bindings.
+// Since they are used in derive macros,
+// it is useful that they are accessible at the root of the crate.
+pub(crate) use js::gc::Traceable as JSTraceable;
